requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
pyyaml
loguru
jupyter
pytest 